fe/src/app/screens/card/card.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { MatDialogModule } from '@angular/material/dialog';
import { MatDialog } from '@angular/material/dialog';
import { CardComponent } from './card.component';

describe('CardComponent', () => {
  let component: CardComponent;
  let fixture: ComponentFixture<CardComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [CardComponent],
      imports: [MatDialogModule],
      providers: [MatDialog],
    }).compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(CardComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  // it('should create', () => {
  //   expect(component).toBeTruthy();
  // });

  it('should return green color when value is greater than 10', () => {
    const value = '15';

    const color = component.getPropColor(value);

    expect(color).toBe('green');
  });

  it('should return inherit color when value is less than or equal to 10', () => {
    const value = '5';

    const color = component.getPropColor(value);

    expect(color).toBe('inherit');
  });
});
